+++ turbinia/pubsub.py
@@ -16,231 +16,18 @@
 
 from __future__ import unicode_literals
 
-import copy
-import json
 import logging
-import Queue
-import uuid
 
 from google.cloud import pubsub
 
-import celery
-import kombu
-from amqp.exceptions import ChannelError
-
 # Turbinia
 from turbinia import config
-from turbinia import evidence
-from turbinia import TurbiniaException
+from turbinia.message import TurbiniaMessageBase
 
 log = logging.getLogger('turbinia')
 
 
-class TurbiniaRequest(object):
-  """An object to request evidence to be processed.
-
-  Attributes:
-    request_id: A client specified ID for this request.
-    recipe: Recipe to use when processing this request.
-    context: A Dict of context data to be passed around with this request.
-    evidence: A list of Evidence objects.
-  """
-
-  def __init__(
-      self, request_id=None, recipe=None, context=None, evidence_=None):
-    """Initialization for TurbiniaRequest."""
-    self.request_id = request_id if request_id else uuid.uuid4().hex
-    self.recipe = recipe
-    self.context = context if context else {}
-    self.evidence = evidence_ if evidence_ else []
-    self.type = self.__class__.__name__
-
-  def to_json(self):
-    """Convert object to JSON.
-
-    Returns:
-      A JSON serialized object.
-    """
-    serializable = copy.deepcopy(self.__dict__)
-    serializable['evidence'] = [x.serialize() for x in serializable['evidence']]
-
-    try:
-      serialized = json.dumps(serializable)
-    except TypeError as e:
-      msg = (
-          'JSON serialization of TurbiniaRequest object {0:s} failed: '
-          '{1:s}'.format(self.type, str(e)))
-      raise TurbiniaException(msg)
-
-    return serialized
-
-  def from_json(self, json_str):
-    """Loads JSON serialized data into self.
-
-    Args:
-      json_str (str): Json serialized TurbiniaRequest object.
-
-    Raises:
-      TurbiniaException: If json can not be loaded, or deserialized object is
-          not of the correct type.
-    """
-    try:
-      obj = json.loads(json_str)
-    except ValueError as e:
-      raise TurbiniaException(
-          'Can not load json from string {0:s}'.format(str(e)))
-
-    if obj.get('type', None) != self.type:
-      raise TurbiniaException(
-          'Deserialized object does not have type of {0:s}'.format(self.type))
-
-    obj['evidence'] = [evidence.evidence_decode(e) for e in obj['evidence']]
-    # pylint: disable=attribute-defined-outside-init
-    self.__dict__ = obj
-
-
-class TurbiniaCelery(object):
-  """Celery app object.
-
-  Attributes:
-    app (Celery): The Celery app itself.
-  """
-
-  def __init__(self):
-    """Celery configurations."""
-    self.app = None
-    self.fexec = None
-
-  def fexec(f, *args, **kwargs):
-    # pylint: disable=no-self-argument,method-hidden
-    """Placeholder function, overwritten with the _fexec closure once setup()
-    is called. Once overwritten, will call the specified function with
-    whichever arguments you pass it.
-
-    Arguments:
-      f (function): an arbitrary function
-      args: any positional arguments to be passsed to the function
-      kwargs: any keyword arguments to be passed to the function
-    """
-    raise NotImplementedError
-
-  def _fexec(self):
-    """Closure used to pass functions to workers. We use this instead of having
-    to annotate each individual TurbiniaTask with the Celery @app.task
-    decorator.
-
-    Returns:
-      function: the fexec() function.
-    """
-    @self.app.task(name='fexec')
-    def fexec(f, *args, **kwargs):
-      """Lets us pass in an arbitrary function without Celery annotations"""
-      return f(*args, **kwargs)
-    return fexec
-
-  def setup(self):
-    """Set up Celery"""
-    config.LoadConfig()
-    self.app = celery.Celery(
-        'turbinia',
-        broker=config.CELERY_BROKER,
-        backend=config.CELERY_BACKEND
-    )
-    self.app.conf.update(
-        task_default_queue=config.INSTANCE_ID,
-        # TODO(ericzinnikas): pickle is not secure, we need to replace it with
-        # the default json serializer, but need to figure out how to register
-        # the TurbiniaTask objects with Celery
-        event_serializer='pickle',
-        result_serializer='pickle',
-        task_serializer='pickle',
-        accept_content=['pickle'],
-        # TODO(ericzinnikas): Without task_acks_late Celery workers will start
-        # on one task and prefetch another (i.e. can result in 1 worker getting
-        # 2 plaso jobs while another worker is free). But enabling this causes
-        # problems with certain Celery brokers (duplicated work).
-        task_acks_late=False,
-        task_track_started=True,
-        worker_concurrency=1,
-        worker_prefetch_multiplier=1,
-    )
-    self.fexec = self._fexec()
-
-
-class TurbiniaKombu(object):
-  """Queue object for receiving evidence messages.
-
-  Attributes:
-    queue (Kombu.SimpleBuffer|Kombu.SimpleQueue): evidence queue.
-  """
-
-  def __init__(self, routing_key):
-    """Kombu config."""
-    self.queue = None
-    self.routing_key = routing_key
-
-  def setup(self):
-    """Set up Kombu SimpleBuffer"""
-    config.LoadConfig()
-    conn = kombu.Connection(config.KOMBU_BROKER)
-    if config.KOMBU_DURABLE:
-      self.queue = conn.SimpleQueue(name=self.routing_key)
-    else:
-      self.queue = conn.SimpleBuffer(name=self.routing_key)
-
-  def check_messages(self):
-    """See if we have any messages in the queue.
-
-    Returns:
-      list[TurbiniaRequest]: all evidence requests.
-    """
-    requests = []
-    while True:
-      try:
-        message = self.queue.get(block=False)
-        request = self._validate_message(message.payload)
-        if request:
-          requests.append(request)
-          if self.queue.queue.durable:
-            message.ack()
-      except Queue.Empty:
-        break
-      except ChannelError:
-        break
-
-    log.debug('Received {0:d} messages'.format(len(requests)))
-    return requests
-
-  def send_message(self, message):
-    """Enqueues a message with Kombu"""
-    data = message.encode('utf-8')
-    self.queue.put(data)
-    log.info('Sent message to queue')
-
-  def send_request(self, request):
-    """Sends a TurbiniaRequest request"""
-    self.send_message(request.to_json())
-
-  def _validate_message(self, message):
-    """Validates pubsub messages and returns them as a new TurbiniaRequest obj.
-
-    Args:
-      message: PubSub message string
-
-    Returns:
-      A TurbiniaRequest object or None if there are decoding failures.
-    """
-    request = TurbiniaRequest()
-    try:
-      request.from_json(message)
-    except TurbiniaException as e:
-      log.error('Error decoding message: {0:s}'.format(str(e)))
-      return None
-
-    return request
-
-
-class TurbiniaPubSub(object):
+class TurbiniaPubSub(TurbiniaMessageBase):
   """PubSub client object for Google Cloud.
 
   Attributes:
@@ -263,24 +50,6 @@ class TurbiniaPubSub(object):
     log.debug('Connecting to PubSub Subscription on {0:s}'.format(
         self.topic_name))
     self.subscription = self.topic.subscription(self.topic_name)
-
-  def _validate_message(self, message):
-    """Validates pubsub messages and returns them as a new TurbiniaRequest obj.
-
-    Args:
-      message: PubSub message string
-
-    Returns:
-      A TurbiniaRequest object or None if there are decoding failures.
-    """
-    request = TurbiniaRequest()
-    try:
-      request.from_json(message)
-    except TurbiniaException as e:
-      log.error('Error decoding pubsub message: {0:s}'.format(str(e)))
-      return None
-
-    return request
 
   def check_messages(self):
     """Checks for pubsub messages.
@@ -320,11 +89,3 @@ class TurbiniaPubSub(object):
     log.info(
         'Published message {0:s} to topic {1:s}'.format(
             msg_id, self.topic_name))
-
-  def send_request(self, request):
-    """Sends a TurbiniaRequest message.
-
-    Args:
-      request: A TurbiniaRequest object.
-    """
-    self.send_message(request.to_json())
